# docs/swagger.yaml
basePath: /api
definitions:
  model.AddressResponse:
    properties:
      city:
        type: string
      country:
        type: string
      created_at:
        type: integer
      id:
        type: string
      postal_code:
        type: string
      province:
        type: string
      street:
        type: string
      updated_at:
        type: integer
    type: object
  model.ContactResponse:
    properties:
      addresses:
        items:
          $ref: '#/definitions/model.AddressResponse'
        type: array
      created_at:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      updated_at:
        type: integer
    type: object
...
  model.CreateContactRequest:
    properties:
      email:
        maxLength: 200
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - first_name
    type: object
...
  model.PageMetadata:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_item:
        type: integer
      total_page:
        type: integer
    type: object
...
  model.UpdateContactRequest:
    properties:
      email:
        maxLength: 200
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - first_name
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for managing users, contacts, and addresses with clean
    architecture principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Golang Clean Architecture API
  version: "1.0"
paths:
  /contacts:
    get:
      consumes:
      - application/json
      description: Search and list contacts for the authenticated user with pagination
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by phone
        in: query
        name: phone
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of contacts with pagination
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.ContactResponse'
                type: array
              paging:
                $ref: '#/definitions/model.PageMetadata'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: Create a new contact for the authenticated user
      parameters:
      - description: Contact creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created contact
          schema:
            properties:
              data:
                $ref: '#/definitions/model.ContactResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new contact
      tags:
      - contacts
  /contacts/{contactId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific contact by ID for the authenticated user
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted contact
          schema:
            properties:
              data:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Contact not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a contact
      tags:
      - contacts
    get:
      consumes:
      - application/json
      description: Get a specific contact by ID for the authenticated user
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Contact details
          schema:
            properties:
              data:
                $ref: '#/definitions/model.ContactResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Contact not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: Update a specific contact by ID for the authenticated user
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Contact update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated contact
          schema:
            properties:
              data:
                $ref: '#/definitions/model.ContactResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Contact not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a contact
      tags:
      - contacts
...
securityDefinitions:
  BearerAuth:
    description: 'API token authentication. Format: your-token-here (without "Bearer"
      prefix)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"